+++ src/test/java/blackjack/domain/participant/PlayersTest.java
@@ -7,6 +7,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import blackjack.domain.card.Hand;
+import java.math.BigDecimal;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -26,8 +27,8 @@ public class PlayersTest {
     @Test
     void createPlayers() {
         // given
-        List<Player> players = List.of(new Player(provideEmptyCards(), "엠제이", 10_000),
-                new Player(provideEmptyCards(), "밍트", 20_000));
+        List<Player> players = List.of(new Player(provideEmptyCards(), "엠제이", BigDecimal.valueOf(10_000)),
+                new Player(provideEmptyCards(), "밍트", BigDecimal.valueOf(20_000)));
 
         // when & then
         Assertions.assertThatCode(() -> new Players(players))
@@ -38,8 +39,8 @@ public class PlayersTest {
     @Test
     void createDuplicatePlayers() {
         // given
-        List<Player> players = List.of(new Player(provideEmptyCards(), "엠제이", 10_000),
-                new Player(provideEmptyCards(), "엠제이", 20_000));
+        List<Player> players = List.of(new Player(provideEmptyCards(), "엠제이", BigDecimal.valueOf(10_000)),
+                new Player(provideEmptyCards(), "엠제이", BigDecimal.valueOf(20_000)));
 
         // when & then
         Assertions.assertThatThrownBy(() -> new Players(players))
@@ -60,9 +61,9 @@ public class PlayersTest {
         // then
         assertAll(
                 () -> assertThat(players.getPlayers().getFirst()).isEqualTo(
-                        new Player(hand.subHand(0, 2), "엠제이", 10_000)),
+                        new Player(hand.subHand(0, 2), "엠제이", BigDecimal.valueOf(10_000))),
                 () -> assertThat(players.getPlayers().get(1)).isEqualTo(
-                        new Player(hand.subHand(2, 4), "밍트", 20_000))
+                        new Player(hand.subHand(2, 4), "밍트", BigDecimal.valueOf(20_000)))
         );
     }
 }
